# [Numpy] FFI: split and svd