--- a/python/mxnet/__init__.py
+++ b/python/mxnet/__init__.py
@@ -108,3 +108,5 @@ from . import _global_var
 
 from . import _api_internal
 from . import api
+from . import container
+from . import ndarray_handle