--- a/python/mxnet/__init__.py
+++ b/python/mxnet/__init__.py
@@ -109,4 +109,3 @@ from . import _global_var
 from . import _api_internal
 from . import api
 from . import container
-from . import ndarray_handle