--- a/python/mxnet/_ffi/_cython/core.pyx
+++ b/python/mxnet/_ffi/_cython/core.pyx
@@ -20,4 +20,5 @@
 include "./base.pxi"
 include "./ndarray.pxi"
 include "./convert.pxi"
+include "./object.pxi"
 include "./function.pxi"