--- a/python/mxnet/container.py
+++ b/python/mxnet/container.py
@@ -14,6 +14,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+# pylint: disable=undefined-variable
 """
 Container data structures.
 Acknowledgement: This file originates from incubator-tvm