--- a/python/mxnet/ndarray_handle.py
+++ b/python/mxnet/ndarray_handle.py
@@ -1,0 +1,29 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+"""
+NDarray Handle
+"""
+from ._ffi.object import Object, register_object
+from ._ffi.function import _init_api
+
+@register_object("MXNet.NDArrayHandle")
+class NDArrayHandle(Object):
+    @property
+    def value(self):
+        return _GetNDArrayHandleValue(self)
+
+_init_api("mxnet.ndarray_handle")