--- a/src/api/operator/numpy/linalg/np_gesvd.cc
+++ b/src/api/operator/numpy/linalg/np_gesvd.cc
@@ -40,7 +40,9 @@ MXNET_REGISTER_API("_npi.svd")
   // outputs
   int num_outputs = 0;
   auto ndoutputs = Invoke(op, &attrs, num_inputs, inputs, &num_outputs, nullptr);
-  *ret = ADT(0, {NDArrayHandle(ndoutputs[0]), NDArrayHandle(ndoutputs[1]), NDArrayHandle(ndoutputs[2])});
+  *ret = ADT(0, {NDArrayHandle(ndoutputs[0]),
+                 NDArrayHandle(ndoutputs[1]),
+                 NDArrayHandle(ndoutputs[2])});
 });
 
 }  // namespace mxnet