--- a/src/api/operator/op_utils.cc
+++ b/src/api/operator/op_utils.cc
@@ -28,7 +28,7 @@
 
 namespace mxnet {
 
-std::string String2MXNetTypeWithBool(int dtype) {
+std::string MXNetTypeWithBool2String(int dtype) {
   switch (dtype) {
     case mshadow::kFloat32:
       return "float32";